requirements.txt
django==3.2.5
ipdb
ipython
psycopg2_binary
requests
whitenoise
gunicorn
django-extensions
django-debug-toolbar==3.2.1
python-slugify
pycco
django-hijack==2.3.0
django-hijack-admin
usabilla-api
user-agents
pandas
apscheduler
django-model-utils
sendgrid
boto3
django-storages
docutils
django-request
django-queryset-csv
ibm-cos-sdk
psutil
